fetch submodule (#106)
* fetch submodule
* fix compile error
* fix for gperftools installation; remove redundant profiler linkage;

+++ test/s3select_test.h
@@ -235,8 +235,8 @@ int run_query_on_parquet_file(const char* input_query, const char* input_file, s
   std::function<size_t(int64_t,int64_t,void*,optional_yield*)> fp_range_req=[&](int64_t start,int64_t length,void *buff,optional_yield*y)
   {
     fseek(fp,start,SEEK_SET);
-    fread(buff, length, 1, fp);
-    return length;
+    size_t read_sz = fread(buff, 1, length, fp);
+    return read_sz;
   };
 
   rgw_s3select_api rgw;
@@ -330,7 +330,12 @@ std::string run_expression_in_C_prog(const char* expression)
     return std::string("#ERROR#");
   }
 
-  fgets(result_buff, sizeof(result_buff), fp_build);
+  char * res = fgets(result_buff, sizeof(result_buff), fp_build);
+
+  if(!res)
+  {
+    return std::string("#ERROR#");
+  }
 
   unlink(c_run_file.c_str());
   unlink(c_test_file.c_str());
